requirements.txt
selenium
fastapi
uvicorn
pandas
requests
psycopg2-binary
sqlalchemy
apache-airflow
python-dotenv